# services/api-server/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ALPA Inventory API",
    "version": "1.0.0",
        "description": "API for maintaining workspaces, products, and components. Input strings are trimmed; programming language values are normalized to lowercase. Business validation returns domain error codes; JSON, length, and character validation returns invalid_request."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
    "/v1/workspaces": {
      "post": {
        "tags": [
          "Workspaces"
        ],
        "summary": "Create a workspace",
        "operationId": "createWorkspace",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workspace created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "get": {
        "tags": [
          "Workspaces"
        ],
        "summary": "List workspaces",
        "operationId": "listWorkspaces",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/WorkspaceID"
        }
      ],
      "get": {
        "tags": [
          "Workspaces"
        ],
        "summary": "Get a workspace",
        "operationId": "getWorkspace",
        "responses": {
          "200": {
            "description": "Workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "tags": [
          "Workspaces"
        ],
        "summary": "Delete a workspace",
        "operationId": "deleteWorkspace",
        "responses": {
          "204": {
            "description": "Workspace deleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/products": {
      "parameters": [
        {
          "$ref": "#/components/parameters/WorkspaceID"
        }
      ],
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create a product in a workspace",
        "operationId": "createProduct",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "List products in a workspace",
        "operationId": "listWorkspaceProducts",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Product page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/products/{product_id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProductID"
        }
      ],
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get a product",
        "operationId": "getProduct",
        "responses": {
          "200": {
            "description": "Product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Delete a product",
        "operationId": "deleteProduct",
        "responses": {
          "204": {
            "description": "Product deleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/components": {
      "post": {
        "tags": [
          "Components"
        ],
        "summary": "Create a component",
        "operationId": "createComponent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateComponentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Component created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Component"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/components/{component_id}/apis": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ParentComponentID"
        }
      ],
      "post": {
        "tags": [
          "APIs"
        ],
        "summary": "Create a component API",
        "operationId": "createComponentAPI",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAPIRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComponentAPI"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/components/{component_id}/apis/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ParentComponentID"
        },
        {
          "$ref": "#/components/parameters/APIID"
        }
      ],
      "delete": {
        "tags": [
          "APIs"
        ],
        "summary": "Delete a component API",
        "operationId": "deleteComponentAPI",
        "responses": {
          "204": {
            "description": "API deleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "put": {
        "tags": [
          "APIs"
        ],
        "summary": "Update a component API",
        "operationId": "updateComponentAPI",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAPIRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComponentAPI"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/components/{component_id}/clients": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ParentComponentID"
        }
      ],
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Create a component client",
        "operationId": "createComponentClient",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClientRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Client created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComponentClient"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/components/{component_id}/clients/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ParentComponentID"
        },
        {
          "$ref": "#/components/parameters/ClientID"
        }
      ],
      "delete": {
        "tags": [
          "Clients"
        ],
        "summary": "Delete a component client",
        "operationId": "deleteComponentClient",
        "responses": {
          "204": {
            "description": "Client deleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "put": {
        "tags": [
          "Clients"
        ],
        "summary": "Update a component client",
        "operationId": "updateComponentClient",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClientRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Client updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComponentClient"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/components/{component_id}/clients/{id}/bindings": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ParentComponentID"
        },
        {
          "$ref": "#/components/parameters/ClientID"
        }
      ],
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Bind a component client to an API",
        "operationId": "createClientBinding",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClientBindingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Client binding created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientBinding"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/components/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ComponentID"
        }
      ],
      "get": {
        "tags": [
          "Components"
        ],
        "summary": "Get a component",
        "operationId": "getComponent",
        "responses": {
          "200": {
            "description": "Component",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Component"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "tags": [
          "Components"
        ],
        "summary": "Delete a component",
        "operationId": "deleteComponent",
        "responses": {
          "204": {
            "description": "Component deleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/products/{product_id}/components": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProductID"
        }
      ],
      "get": {
        "tags": [
          "Components"
        ],
        "summary": "List components in a product",
        "operationId": "listProductComponents",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Component page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComponentList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "WorkspaceID": {
        "name": "workspace_id",
        "in": "path",
        "required": true,
        "schema": {
          "allOf": [
            {
              "$ref": "#/components/schemas/ID"
            }
          ],
          "example": 1001
        }
      },
      "ProductID": {
        "name": "product_id",
        "in": "path",
        "required": true,
        "schema": {
          "allOf": [
            {
              "$ref": "#/components/schemas/ID"
            }
          ],
          "example": 2001
        }
      },
...
      "ParentComponentID": {
        "name": "component_id",
        "in": "path",
        "required": true,
        "schema": {
          "allOf": [
            {
              "$ref": "#/components/schemas/ID"
            }
          ],
          "example": 3001
        }
      },
      "APIID": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "allOf": [
            {
              "$ref": "#/components/schemas/ID"
            }
          ],
          "example": 4001
        }
      },
      "ClientID": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "allOf": [
            {
              "$ref": "#/components/schemas/ID"
            }
          ],
          "example": 5001
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20,
          "example": 20
        }
      },
      "Offset": {
        "name": "offset",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0,
          "example": 0
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflicting resource state",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Workspace": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ID"
              }
            ],
            "example": 1001
          },
          "name": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Name"
              }
            ],
            "example": "Payments Platform"
          }
        }
      },
      "CreateWorkspaceRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Name"
              }
            ],
            "example": "Payments Platform"
          }
        }
      },
      "WorkspaceList": {
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Workspace"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
...
      "Product": {
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "owning_team_id",
          "product_code",
          "name",
          "criticality",
          "description"
        ],
        "properties": {
          "id": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ID"
              }
            ],
            "example": 2001
          },
          "workspace_id": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ID"
              }
            ],
            "example": 1001
          },
          "owning_team_id": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ID"
              }
            ],
            "nullable": true,
            "example": 301
          },
          "product_code": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductCode"
              }
            ],
            "example": "CHECKOUT"
          },
          "name": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Name"
              }
            ],
            "example": "Checkout Service"
          },
          "criticality": {
            "$ref": "#/components/schemas/Criticality"
          },
          "description": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Description"
              }
            ],
            "nullable": true,
            "example": "Handles checkout and payment confirmation."
          }
        }
      },
      "CreateProductRequest": {
        "type": "object",
        "required": [
          "product_code",
          "name",
          "criticality"
        ],
        "properties": {
          "owning_team_id": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ID"
              }
            ],
            "example": 301
          },
          "product_code": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductCode"
              }
            ],
            "example": "CHECKOUT"
          },
          "name": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Name"
              }
            ],
            "example": "Checkout Service"
          },
          "criticality": {
            "$ref": "#/components/schemas/Criticality"
          },
          "description": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Description"
              }
            ],
            "example": "Handles checkout and payment confirmation."
          }
        }
      },
      "ProductList": {
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
...
      "Component": {
        "type": "object",
        "required": [
          "id",
          "product_id",
          "name",
          "type",
          "description",
          "details",
          "apis",
          "clients"
        ],
        "properties": {
          "id": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ID"
              }
            ],
            "example": 3001
          },
          "product_id": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ID"
              }
            ],
            "example": 2001
          },
          "name": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Name"
              }
            ],
            "example": "Checkout API"
          },
          "type": {
            "$ref": "#/components/schemas/ComponentType"
          },
          "description": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Description"
              }
            ],
            "example": "Serves checkout requests for the web and mobile apps."
          },
          "details": {
            "$ref": "#/components/schemas/ComponentDetails"
          },
          "apis": {
            "type": "array",
            "maxItems": 5,
            "items": {
              "$ref": "#/components/schemas/ComponentAPI"
            }
          },
          "clients": {
            "type": "array",
            "maxItems": 5,
            "items": {
              "$ref": "#/components/schemas/ComponentClient"
            }
          }
        }
      },
      "CreateComponentRequest": {
        "type": "object",
        "required": [
          "product_id",
          "name",
          "type",
          "details"
        ],
        "properties": {
          "product_id": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ID"
              }
            ],
            "example": 2001
          },
          "name": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Name"
              }
            ],
            "example": "Checkout API"
          },
          "type": {
            "$ref": "#/components/schemas/ComponentType"
          },
          "description": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Description"
              }
            ],
            "example": "Serves checkout requests for the web and mobile apps."
          },
          "details": {
            "$ref": "#/components/schemas/ComponentDetailsRequest"
          },
          "apis": {
            "type": "array",
            "maxItems": 5,
            "items": {
              "$ref": "#/components/schemas/CreateAPIRequest"
            }
          },
          "clients": {
            "type": "array",
            "maxItems": 5,
            "items": {
              "$ref": "#/components/schemas/CreateClientRequest"
            }
          }
        }
      },
...
      "ComponentAPI": {
        "type": "object",
        "required": [
          "id",
          "name",
          "api_type",
          "network_exposure",
          "documentation_url"
        ],
        "properties": {
          "id": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ID"
              }
            ],
            "example": 4001
          },
          "name": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Name"
              }
            ],
            "example": "Checkout Public API"
          },
          "api_type": {
            "$ref": "#/components/schemas/APIType"
          },
          "network_exposure": {
            "$ref": "#/components/schemas/NetworkExposure"
          },
          "documentation_url": {
            "type": "string",
            "nullable": true,
            "format": "uri",
            "example": "https://docs.example.com/checkout"
          }
        }
      },
      "CreateAPIRequest": {
        "type": "object",
        "required": [
          "name",
          "api_type",
          "network_exposure"
        ],
        "additionalProperties": false,
        "properties": {
          "name": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Name"
              }
            ],
            "example": "Checkout Public API"
          },
          "api_type": {
            "$ref": "#/components/schemas/APIType"
          },
          "network_exposure": {
            "$ref": "#/components/schemas/NetworkExposure"
          },
          "documentation_url": {
            "type": "string",
            "nullable": true,
            "format": "uri",
            "example": "https://docs.example.com/checkout"
          }
        }
      },
      "UpdateAPIRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CreateAPIRequest"
          }
        ]
      },
...
      "CreateClientRequest": {
        "type": "object",
        "required": [
          "client_name",
          "role"
        ],
        "additionalProperties": false,
        "properties": {
          "client_name": {
            "$ref": "#/components/schemas/ClientName"
          },
          "role": {
            "$ref": "#/components/schemas/ClientRole"
          },
          "action": {
            "type": "string",
            "nullable": true,
            "maxLength": 2097152,
            "example": "# Call checkout"
          },
          "capabilities": {
            "type": "string",
            "nullable": true,
            "maxLength": 2097152,
            "example": "read:orders"
          },
          "secure_connection": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "UpdateClientRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CreateClientRequest"
          }
        ]
      },
...
      "CreateClientBindingRequest": {
        "type": "object",
        "required": [
          "api_id"
        ],
        "additionalProperties": false,
        "properties": {
          "api_id": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ID"
              }
            ],
            "example": 4001
          }
        }
      }
    }
  }
}