# services/api-server/internal/httpapi/routing.go
package httpapi

import (
	"net/http"
	"strings"

	"github.com/76parker/alpa/docs"
	"github.com/76parker/alpa/internal/httpapi/apis"
	"github.com/76parker/alpa/internal/httpapi/clients"
	"github.com/76parker/alpa/internal/httpapi/component"
	"github.com/76parker/alpa/internal/httpapi/product"
	"github.com/76parker/alpa/internal/httpapi/workspace"
	"github.com/76parker/alpa/internal/observability/logger"
	"github.com/gin-gonic/gin"
	swaggerFiles "github.com/swaggo/files/v2"
)

type Handlers struct {
	UI        http.Handler
	Workspace *workspace.Handler
	Product   *product.Handler
	Component *component.Handler
	APIs      *apis.Handler
	Clients   *clients.Handler
}

func newRouter(log logger.Logger, handlers Handlers) *gin.Engine {
	gin.SetMode(gin.ReleaseMode)
	router := gin.New()
	router.Use(withLogging(log))
	initSwaggerRoutes(router)
	router.Use(ErrorHandler())

	v1 := router.Group("/v1")
	initWorkspaceRoutes(v1, &handlers)
	initProductRoutes(v1, &handlers)
	initComponentRoutes(v1, &handlers)
	initAPIRoutes(v1, &handlers)
	initClientRoutes(v1, &handlers)
	if handlers.UI != nil {
		router.NoRoute(func(c *gin.Context) {
			first := strings.SplitN(strings.TrimPrefix(c.Request.URL.Path, "/"), "/", 2)[0]
			if first == "v1" || first == "docs" || first == "api" {
				c.Status(http.StatusNotFound)
				return
			}
			c.Set("handler_name", "webui")
			handlers.UI.ServeHTTP(c.Writer, c.Request)
		})
	}
	return router
}

func serveSwaggerUI(c *gin.Context) {
	c.Data(http.StatusOK, "text/html; charset=utf-8", []byte(docs.SwaggerUIHTML))
}

func serveOpenAPIYAML(c *gin.Context) {
	c.Data(http.StatusOK, "application/yaml; charset=utf-8", docs.OpenAPIYAML)
}

func serveOpenAPIJSON(c *gin.Context) {
	c.Data(http.StatusOK, "application/json; charset=utf-8", docs.OpenAPIJSON)
}

func initSwaggerRoutes(router *gin.Engine) {
	router.GET("/docs", handlerName("documentation"), serveSwaggerUI)
	router.GET("/docs/openapi.yaml", handlerName("documentation"), serveOpenAPIYAML)
	router.GET("/docs/openapi.json", handlerName("documentation"), serveOpenAPIJSON)
	router.GET(
		"/docs/swagger-ui/*filepath",
		handlerName("documentation"),
		gin.WrapH(http.StripPrefix("/docs/swagger-ui/", http.FileServer(http.FS(swaggerFiles.FS)))),
	)
}

func initComponentRoutes(group *gin.RouterGroup, handlers *Handlers) {
	group.POST("/components", handlerName("component.create"), handlers.Component.Create)
	group.GET("/components/:component_id", handlerName("component.get"), handlers.Component.Get)
	group.DELETE("/components/:component_id", handlerName("component.delete"), handlers.Component.Delete)
	group.GET("/products/:product_id/components", handlerName("component.list"), handlers.Component.List)
}

func initAPIRoutes(group *gin.RouterGroup, handlers *Handlers) {
	group.POST("/components/:component_id/apis", handlerName("api.create"), handlers.APIs.Create)
	group.PUT("/components/:component_id/apis/:id", handlerName("api.update"), handlers.APIs.Update)
	group.DELETE("/components/:component_id/apis/:id", handlerName("api.delete"), handlers.APIs.Delete)
}

func initClientRoutes(group *gin.RouterGroup, handlers *Handlers) {
	group.POST("/components/:component_id/clients", handlerName("client.create"), handlers.Clients.Create)
	group.PUT("/components/:component_id/clients/:id", handlerName("client.update"), handlers.Clients.Update)
	group.DELETE("/components/:component_id/clients/:id", handlerName("client.delete"), handlers.Clients.Delete)
	group.POST(
		"/components/:component_id/clients/:id/bindings",
		handlerName("client_binding.create"),
		handlers.Clients.BindAPI,
	)
}

func initProductRoutes(group *gin.RouterGroup, handlers *Handlers) {
	group.POST("/workspaces/:workspace_id/products", handlerName("product.create"), handlers.Product.Create)
	group.GET("/workspaces/:workspace_id/products", handlerName("product.list"), handlers.Product.List)
	group.GET("/products/:product_id", handlerName("product.get"), handlers.Product.Get)
	group.DELETE("/products/:product_id", handlerName("product.delete"), handlers.Product.Delete)
}

func initWorkspaceRoutes(group *gin.RouterGroup, handlers *Handlers) {
	group.POST("/workspaces", handlerName("workspace.create"), handlers.Workspace.Create)
	group.GET("/workspaces", handlerName("workspace.list"), handlers.Workspace.List)
	group.GET("/workspaces/:workspace_id", handlerName("workspace.get"), handlers.Workspace.Get)
}